chore: move to monorepo architecture with pnpm
