+++ integration/package.json
@@ -1,0 +1,14 @@
+{
+  "private": "true",
+  "dependencies": {
+    "nestjs-commander": "workspace:*",
+    "nestjs-commander-testing": "workspace:*",
+    "commander": "^6.2.1",
+    "@nestjs/common": "^7.6.5",
+    "@nestjs/core": "^7.6.5",
+    "@nestjs/testing": "^7.6.5",
+    "jest": "^26.6.3",
+    "typescript": "^4.1.3",
+    "ts-jest": "^26.4.4"
+  }
+}
