+++ jest.integration.config.js
@@ -1,0 +1,6 @@
+const baseConfig = require('./jest.config');
+module.exports = {
+  ...baseConfig,
+  testRegex: 'integration/.*.spec.ts$',
+  collectCoverage: false,
+};
