+++ package.json
@@ -1,13 +1,16 @@
 {
-  "name": "nestjs-commander",
+  "name": "nestjs-commander-monorepo",
   "version": "1.0.0",
   "description": "A module for making CLI applications with NestJS. Decorators for running commands and separating out config parsers included. This package works on top of commander.",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
-    "build": "nest build",
-    "test": "jest -c jest.config.js test"
+    "test": "jest -c jest.config.js test",
+    "lint": "eslint --ext .ts .",
+    "format": "prettier --write \"{packages,integration}/**/{src,test}/*.ts\"",
+    "test:int": "jest -c jest.integration.config.js"
   },
+  "private": true,
   "repository": {
     "type": "git",
     "url": "git+https://github.com/jmcdo29/nestjs-commander.git"
@@ -46,16 +49,28 @@
     "lint-staged": "^10.5.3",
     "prettier": "^2.2.1",
     "reflect-metadata": "^0.1.13",
+    "rxjs": "^6.6.3",
     "ts-jest": "^26.4.4",
-    "ts-loader": "^8.0.14",
     "typescript": "^4.1.3"
   },
-  "dependencies": {
-    "@golevelup/nestjs-discovery": "^2.3.1",
-    "commander": "^6.2.1"
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged",
+      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
+    }
   },
-  "peerDependencies": {
-    "@nestjs/common": "^7.0.0",
-    "@nestjs/core": "^7.0.0"
+  "lint-staged": {
+    "*.ts": [
+      "prettier --write",
+      "eslint --ext ts"
+    ],
+    "*.{md,html,json,js}": [
+      "prettier --write"
+    ]
+  },
+  "config": {
+    "commitizen": {
+      "path": "./node_modules/cz-conventional-changelog"
+    }
   }
 }
