+++ packages/nestjs-commander-testing/tsconfig.build.json
@@ -1,0 +1,8 @@
+{
+  "extends": "./../../tsconfig.build.json",
+  "compilerOptions": {
+    "rootDir": "./src",
+    "outDir": "./dist"
+  },
+  "exclude": ["test"]
+}
