+++ packages/nestjs-commander/README.md
@@ -65,7 +65,7 @@ And that's it. Under the hood, `CommandFactory` will worry about calling `NestFa
 
 ## Testing
 
-So what's the use of writing a super awesome command line script if you can't test it super easily, right? Fortunately, `nestjs-commander` has some utilities you can make use of that fits in perfectly with the NestJS ecosystem, it'll feel right at home to any Nestlings out there. Instead of using the `CommandFactory` for building the command in test mode, you can use `CommandTestFactory` and pass in your metadata, very similarly to how `Test.createTestingModule` from `@nestjs/testing` works. In fact, it uses this package under the hood. You're also still able to chain on the `overrideProvider` methods before calling `compile()` so you can swap out DI pieces right in the test. [A nice example of this can be seen in the basic.command.factory.spec.ts file](./integration/test/basic.command.factory.spec.ts).
+There is a testing helper package called [`nestjs-commander-testing`](./../nestjs-commander-testing/README.md) that works very similarly to `@nestjs/testing`. Check out it's documentation and examples for help.
 
 ## Putting it All Together
