+++ packages/nestjs-commander/package.json
@@ -1,0 +1,56 @@
+{
+  "name": "nestjs-commander",
+  "version": "1.0.0",
+  "description": "A module for making CLI applications with NestJS. Decorators for running commands and separating out config parsers included. This package works on top of commander.",
+  "scripts": {
+    "prebuild": "rm -rf dist",
+    "build": "nest build",
+    "test": "jest -c jest.config.js test"
+  },
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "publishConfig": {
+    "main": "dist/index.js",
+    "typings": "dist/index.d.ts"
+  },
+  "files": [
+    "dist"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/jmcdo29/nestjs-commander.git"
+  },
+  "keywords": [
+    "cli",
+    "nestjs",
+    "application",
+    "command",
+    "command-line",
+    "nest",
+    "decorator"
+  ],
+  "author": "Jay McDoniel <jmcdo29@gmail.com>",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/jmcdo29/nestjs-commander/issues"
+  },
+  "homepage": "https://github.com/jmcdo29/nestjs-commander#readme",
+  "devDependencies": {
+    "@nestjs/cli": "^7.5.4",
+    "@nestjs/common": "^7.6.5",
+    "@nestjs/core": "^7.6.5",
+    "@nestjs/testing": "^7.6.5",
+    "jest": "^26.6.3",
+    "reflect-metadata": "^0.1.13",
+    "ts-jest": "^26.4.4",
+    "typescript": "^4.1.3"
+  },
+  "dependencies": {
+    "@golevelup/nestjs-discovery": "^2.3.1",
+    "commander": "^6.2.1"
+  },
+  "peerDependencies": {
+    "@nestjs/common": "^7.0.0",
+    "@nestjs/core": "^7.0.0"
+  }
+}
