+++ packages/nestjs-commander/src/index.ts
@@ -1,4 +1,5 @@
 export * from './command-runner.interface';
-export * from './command-test.factory';
+export * from './command-runner-core.module';
+export * from './command-runner-core.service';
 export * from './command.factory';
 export * from './command.decorators';
